refactor: Activity tab
