+++ apps/meteor/client/views/omnichannel/contactInfo/tabs/ContactInfoActivity/components/index.ts
@@ -1,0 +1,1 @@
+export { default } from './ActivityListItem';
