+++ apps/meteor/client/views/omnichannel/contactInfo/tabs/ContactInfoActivity/useActivityList.tsx
@@ -1,22 +1,19 @@
 import { useQuery } from '@tanstack/react-query';
-import type { ComponentProps } from 'react';
 
-import type ContactInfoActivityItem from './ActivityListItem';
+import { mockActivityList } from './mocks';
+
+type ActivityStatus = 'sent' | 'delivered' | 'read' | 'failed';
+
+export type ActivityItem = {
+	id: string;
+	name: string;
+	status: { id: ActivityStatus; ts: string };
+};
 
 export const useActivityList = () => {
 	// const getActivities = useEndpoint('GET', '/v1/omnichannel/activities');
-	return useQuery<ComponentProps<typeof ContactInfoActivityItem>[]>({
+	return useQuery<ActivityItem[]>({
 		queryKey: ['/v1/omnichannel/activities'],
-		queryFn: () =>
-			new Promise((res) =>
-				setTimeout(
-					() =>
-						res([
-							{ name: 'Haylie George', status: { id: 'sent', ts: new Date().toISOString() } },
-							{ name: 'Haylie George', status: { id: 'sent', ts: new Date().toISOString() } },
-						]),
-					2000,
-				),
-			),
+		queryFn: () => new Promise((res) => setTimeout(() => res(mockActivityList), 2000)),
 	});
 };
